"COMPLETE - PROCESO DEL PACIENTE"

+++ accounts/views.py
@@ -1,3 +1,4 @@
+from django.db import transaction
 from rest_framework.response import Response
 from rest_framework.views import APIView
 from rest_framework_simplejwt.tokens import RefreshToken
@@ -7,8 +8,10 @@ from rest_framework_simplejwt.authentication import JWTAuthentication
 from django.contrib.auth.models import User
 from django.contrib.auth import authenticate
 from rest_framework import status
-from .serializers import UserRegistrationSerializer, UserLoginSerializer, UserAndPatientRegistrationSerializer, ChangePasswordSerializer
-from patients.serializers import PatientSerializer
+from .serializers import UserSerializer, UserRegisterSerializer, UserLoginSerializer, ChangePasswordSerializer
+from patients.serializers import ClientPatientSerializer, PatientSerializer
+from employees.serializers import EmployeeSerializer
+from medical_records.serializers import MedicalRecordSerializer
 
 # EMAIL
 from django.contrib.auth.tokens import default_token_generator
@@ -19,13 +22,23 @@ from django.core.mail import send_mail
 from django.shortcuts import render, redirect
 from django.contrib.auth.forms import SetPasswordForm
 
+# Respuesta
+def create_response(success, message, data=None, errors=None, status_code=status.HTTP_200_OK):
+    return Response({
+        'success': success,
+        'message': message,
+        'data': data,
+        'errors': errors,
+    }, status=status_code)
+
+# RESET PASSWOR EMAIL (COMPLETE)
 class PasswordResetRequestAPIView(APIView):
-    permission_classes = (AllowAny,)
+    permission_classes = (IsAuthenticated,)
 
     def post(self, request):
         email = request.data.get('email')
         if not email:
-            return Response({'error': 'Gmail requerido'}, status=status.HTTP_400_BAD_REQUEST)
+            return create_response(False, 'Email requerido', status_code=status.HTTP_400_BAD_REQUEST)
 
         user = User.objects.filter(email=email).first()
         if user:
@@ -39,10 +52,11 @@ class PasswordResetRequestAPIView(APIView):
             })
             send_mail(mail_subject, message, 'joseleo042001@gmail.com', [email])
 
-        return Response({'message': 'Si existe una cuenta con ese correo, recibirá un correo con instrucciones.'}, status=status.HTTP_200_OK)
+        return create_response(True, 'Si existe una cuenta con ese correo, recibirá un correo con instrucciones.')
 
+# CONFIRM PASSWORD (COMPLETE)
 class PasswordResetConfirmAPIView(APIView):
-    permission_classes = (AllowAny,)
+    permission_classes = (IsAuthenticated,)
 
     def post(self, request):
         uidb64 = request.data.get('uid')
@@ -58,123 +72,73 @@ class PasswordResetConfirmAPIView(APIView):
         if user and default_token_generator.check_token(user, token):
             user.set_password(new_password)
             user.save()
-            return Response({'message': 'La contraseña se ha restablecido correctamente'}, status=status.HTTP_200_OK)
+            return create_response(True, 'La contraseña se ha restablecido correctamente', status_code=status.HTTP_200_OK)
         else:
-            return Response({'error': 'Token o ID de usuario no válido'}, status=status.HTTP_400_BAD_REQUEST)
+            return create_response(False, 'Token o ID de usuario no válido', status_code=status.HTTP_400_BAD_REQUEST)
 
-"""class PasswordResetConfirmAPIView(APIView):
+#REGISTRO DE CUENTA PARA PACIENTE-CLIENTE (COMPLETE)
+class UserPatientRegisterAPIView(APIView):
     permission_classes = (AllowAny,)
-
-    def get(self, request, uidb64, token):
-        try:
-            uid = urlsafe_base64_decode(uidb64).decode()
-            user = User.objects.get(pk=uid)
-        except (TypeError, ValueError, OverflowError, User.DoesNotExist):
-            user = None
-
-        if user and default_token_generator.check_token(user, token):
-            return render(request, 'password_reset_confirm.html', {'uid': uid, 'token': token})
-        else:
-            return render(request, 'password_reset_invalid.html')
+    serializer_class = UserRegisterSerializer
     
-    def post(self, request, uidb64, token):
-        try:
-            uid = urlsafe_base64_decode(uidb64).decode()
-            user = User.objects.get(pk=uid)
-        except (TypeError, ValueError, OverflowError, User.DoesNotExist):
-            user = None
-
-        if user and default_token_generator.check_token(user, token):
-            form = SetPasswordForm(user, request.POST)
-            if form.is_valid():
-                form.save()
-                return redirect('password_reset_complete')
-            else:
-                return render(request, 'password_reset_confirm.html', {'form': form})
-        else:
-            return render(request, 'password_reset_invalid.html')"""
-
-class CustomerRegistrationAPIView(APIView):
-    serializer_class = UserAndPatientRegistrationSerializer
-    permission_classes = (AllowAny,)
-
-    def get(self, request):
-        content = {'message': 'Bienvenido!'}
-        return Response(content)
-
+    @transaction.atomic
     def post(self, request):
+        #Registro de usuario
         serializer = self.serializer_class(data=request.data)
-        if serializer.is_valid(raise_exception=True):
-            new_user = serializer.save()
-            if new_user:
-                refresh = RefreshToken.for_user(new_user)
-                data = {
-                    'refresh': str(refresh),
-                    'access': str(refresh.access_token),
-                }
-                response = Response(data, status=status.HTTP_201_CREATED)
-                response.set_cookie(key='access_token', value=data['access'], httponly=True, secure=True)
-                return response
-        return Response(serializer.errors, status=status.HTTP_400_BAD_REQUEST)
-
-class UserRegistrationAPIView(APIView):
-    serializer_class = UserRegistrationSerializer
-    authentication_classes = (JWTAuthentication,)
-    permission_classes = (IsAuthenticated, IsAdminUser)
-
-    def post(self, request):
-        user = request.user
-        if not user.is_staff:
-            return Response({'message': 'No tienes permiso para realizar esta acción.'}, status=status.HTTP_403_FORBIDDEN)
-
-        serializer = self.serializer_class(data=request.data)
-        if serializer.is_valid(raise_exception=True):
-            new_user = serializer.save()
-            data = {'message': 'Usuario registrado exitosamente'}
-            return Response(data, status=status.HTTP_201_CREATED)
-        return Response(serializer.errors, status=status.HTTP_400_BAD_REQUEST)
-
-#LOGIN
-class UserLoginAPIView(APIView):
-    permission_classes = (AllowAny,)
-    
-    def post(self, request):
-        user_name = request.data.get('username', None)
-        user_password = request.data.get('password', None)
-        user_type = request.data.get('user_type', None)
-        if not user_name or not user_password or not user_type:
-            return Response({'error': 'Faltan campos obligatorios'}, status=400)
+        serializer.is_valid(raise_exception=True)
+        user = serializer.save()
         
-        user_instance = authenticate(username=user_name, password=user_password)
+        #Registro de paciente
+        patient_data = request.data.get('patient')
+        patient_data['user'] = user.id
+        patient_serializer = ClientPatientSerializer(data=patient_data)
+        patient_serializer.is_valid(raise_exception=True)
+        patient = patient_serializer.save()
         
-        if not user_instance:
-            return Response({'error': 'Credenciales inválidas'}, status=403)
-
-        if not user_instance.is_active:
-            return Response({'error': 'Usuario inactivo'}, status=403)
-
-        if user_type == 'P' and not user_instance.is_staff:
-            return Response({'error': 'Acceso denegado, no es parte del personal'}, status=403)
-        elif user_type == 'C' and user_instance.is_staff:
-            #worker = user_instance.worker
-            ##person = WorkerSerializer(worker).data
-            return Response({'error': 'Acceso denegado, no es parte del cliente'}, status=403)
-
-        if not user_instance.is_staff:
-            patient = user_instance.patient
-            person = PatientSerializer(patient).data
-        else:
-            #worker = user_instance.worker
-            person = None
-            
-        refresh = RefreshToken.for_user(user_instance)
-        response = Response({
-            'succes': 'Bienvenido',
-            'person': person,
-            'user': UserLoginSerializer(user_instance).data,
+        # Registro de Historia clinica
+        medical_record_data = {'patient': patient.id}
+        medical_history_serializer = MedicalRecordSerializer(data=medical_record_data)
+        medical_history_serializer.is_valid(raise_exception=True)
+        medical_history_serializer.save()
+        
+        # Tokens de acceso
+        refresh = RefreshToken.for_user(user)
+        data = {
             'refresh': str(refresh),
             'access': str(refresh.access_token),
-        }, status=200)
+        }
+        response = create_response(True, 'Usuario registrado exitosamente', data=data, status_code=status.HTTP_201_CREATED)
+        response.set_cookie(key='access_token', value=data['access'], httponly=True, secure=True)
+        return response
+
+#LOGIN (COMPLETE)
+class UserLoginAPIView(APIView):
+    permission_classes = (AllowAny,)
+    serializer_class = UserLoginSerializer
+    
+    def post(self, request):
+        serializer = self.serializer_class(data=request.data)
+        serializer.is_valid(raise_exception=True)
+        user_instance = authenticate(username=request.data.get('username'), password=request.data.get('password'))
+        # VALIDACIONES
+        if not user_instance:
+            return create_response(False, 'Credenciales inválidas', status_code=status.HTTP_403_FORBIDDEN)
+        if not user_instance.is_active:
+            return create_response(False, 'Usuario inactivo', status_code=status.HTTP_403_FORBIDDEN)
+        
+        if request.data.get('type') == 'P' and not user_instance.is_staff:
+            return create_response(False, 'Acceso denegado, no es parte del personal', status_code=status.HTTP_403_FORBIDDEN)
+        elif request.data.get('type') == 'C' and user_instance.is_staff:
+            #worker = user_instance.worker
+            ##person = WorkerSerializer(worker).data
+            return create_response(False, 'Acceso denegado, no es parte del cliente', status_code=status.HTTP_403_FORBIDDEN)
+            
+        refresh = RefreshToken.for_user(user_instance)
+        response_data = {
+            'refresh': str(refresh),
+            'access': str(refresh.access_token),
+        }
+        response = create_response(True, 'Bienvenido', data=response_data)
         response.set_cookie(
             key='access_token', 
             value=str(refresh.access_token), 
@@ -184,7 +148,7 @@ class UserLoginAPIView(APIView):
         )
         return response
 
-
+#PERFIL (COMPLETE)
 class UserViewAPI(APIView):
     authentication_classes = (JWTAuthentication,)
     permission_classes = (IsAuthenticated,)
@@ -195,25 +159,41 @@ class UserViewAPI(APIView):
             patient = user.patient
             person = PatientSerializer(patient).data
         else:
-            #worker = user_instance.worker
-            person = None
-        return Response({
-            'succes': 'Profile',
+            employee = user.employee
+            person = EmployeeSerializer(employee).data
+        
+        response_data = {
             'person': person,
-            'user': UserLoginSerializer(user).data,
-        }, status=200)
+            'user': UserSerializer(user).data,
+        }
+        return create_response(True, 'Perfil', data=response_data)
 
-
+# CERRAR SESION (COMPLETE - REVOQUE TOKEN REFRESH)
 class UserLogoutViewAPI(APIView):
     authentication_classes = (JWTAuthentication,)
     permission_classes = (IsAuthenticated,)
 
     def post(self, request):
-        response = Response()
-        response.delete_cookie('access_token')
-        response.data = {'message': 'Cerró sesión exitosamente.'}
-        return response
-    
+        try:
+            # Obtener el token de refresco del usuario
+            refresh_token = request.data.get('refresh_token')
+            if not refresh_token:
+                return create_response(False, 'Se requiere token de actualización.', status_code=status.HTTP_400_BAD_REQUEST)
+
+            # Invalidar el token de refresco
+            token = RefreshToken(refresh_token)
+            token.blacklist()
+
+            # Eliminacion de cookie
+            response = create_response(True, 'Se ha cerrado sesión correctamente.', status_code=status.HTTP_200_OK)
+            response.delete_cookie('access_token')
+
+            return response
+
+        except Exception as e:
+            return create_response(False, str(e), status_code=status.HTTP_400_BAD_REQUEST)
+
+# RESTABLECER CONTRASEÑA (COMPLETE)
 class ChangePasswordView(APIView):
     permission_classes = (IsAuthenticated,)
 
@@ -225,11 +205,11 @@ class ChangePasswordView(APIView):
             # Check old password
             old_password = serializer.validated_data['old_password']
             if not user.check_password(old_password):
-                return Response({'old_password': 'Contraseña incorrecta.'}, status=status.HTTP_400_BAD_REQUEST)
+                return create_response(False, 'Contraseña anterior incorrecta.', status_code=status.HTTP_400_BAD_REQUEST)
             
             # Set new password
             user.set_password(serializer.validated_data['new_password'])
             user.save()
-            return Response({'message': 'Contraseña actualizada exitosamente'}, status=status.HTTP_200_OK)
+            return create_response(True, 'Contraseña actualizada exitosamente', status_code=status.HTTP_200_OK)
         
-        return Response(serializer.errors, status=status.HTTP_400_BAD_REQUEST)
+        return create_response(False, 'Error al actualizar la contraseña', errors=serializer.errors, status_code=status.HTTP_400_BAD_REQUEST)
